Fixed tests (15/15)
