+++ pipeline/test_transform.py
@@ -5,25 +5,11 @@
 import pytest
 from pytest import mark
 from unittest.mock import patch, mock_open
-from transform import load_data, clean_dataframe
+from transform import clean_dataframe, save_dataframe_to_csv, summarise_day
 import pandas as pd
 
 
-def test_transform_load_data_type_invalid():
-    """Checks for an invalid type of file_path in load_data."""
-    with pytest.raises(TypeError) as exc:
-        load_data(23)
-    assert str(exc.value) == 'Invalid URL type.'
-
-
-def test_transform_load_data_type_not_csv():
-    """Checks if the file_path ends with a .csv."""
-    with pytest.raises(ValueError) as exc:
-        load_data("data/normalised_output")
-    assert str(exc.value) == 'Please end your filename in .csv.'
-
-
-@patch('transform.load_data')
+@patch('transform.load_csv_data')
 def test_transform_clean_dataframe_valid(test_load_data):
     """Tests data to make sure it is validated"""
 
@@ -33,8 +19,8 @@ def test_transform_clean_dataframe_valid(test_load_data):
     cleaned_dataframe = clean_dataframe('fake_file_path')
 
     # Check for correct column names
-    expected_columns = ["recording_taken", "plant_id", "temperature",
-                        "soil_moisture", "last_watered"]
+    expected_columns = ["temperature", "soil_moisture", "recording_taken",
+                        "last_watered", "plant_id"]
     assert list(cleaned_dataframe.columns) == expected_columns
 
     # Check for no null values
